# docs/openapi.yaml
openapi: 3.0.1
info:
  contact:
    email: mail@heri.life
    name: Open Data Portal
  description: The Data Set API (DSAPI) allows the public users to discover and search
    USPTO exported data sets. This is a generic API that allows USPTO users to make
    any CSV based data files searchable through API. With the help of GET call, it
    returns the list of data fields that are searchable. With the help of POST call,
    data can be fetched based on the filters on the field names. Please note that
    POST call is used to search the actual data. The reason for the POST call is that
    it allows users to specify any complex search criteria without worry about the
    GET size limitations as well as encoding of the input parameters.
  title: DEVCODE BE API
  version: 1.0.0
servers:
- url: http://0.0.0.0:3030
  variables:
    scheme:
      default: https
      description: The Data Set API is accessible via https and http
      enum:
      - https
      - http
tags:
- description: Find out about the data sets
  name: metadata
- description: Search a data set
  name: search
- name: ACTIVITY
- name: TODO
paths:
  /activity-groups:
    get:
      responses:
        "200":
          content:
            application/json:
              example:
                status: Success
                message: Success
                data:
                  schema:
                    $ref: '#/components/schemas/ActivityGroup'
          description: Get activity group response
      summary: Get All
      tags:
      - ACTIVITY
    post:
      requestBody:
        content:
          application/json:
            schema:
              example:
                email: wow@gmail.com
              type: object
      responses:
        "201":
          content:
            application/json:
              example:
                status: Success
                message: Success
                data:
                  schema:
                    $ref: '#/components/schemas/ActivityGroup'
          description: Created activity group response
          headers:
            X-Powered-By:
              explode: false
              schema:
                example: Express
                type: string
              style: simple
            Access-Control-Allow-Origin:
              explode: false
              schema:
                example: '*'
                type: string
              style: simple
            Content-Type:
              explode: false
              schema:
                example: application/json; charset=utf-8
                type: string
              style: simple
            Content-Length:
              explode: false
              schema:
                example: 178
                type: integer
              style: simple
            ETag:
              explode: false
              schema:
                example: W/"b2-5FmzsgzMUXilMXvB5sdXuyL6wNI"
                type: string
              style: simple
            Date:
              explode: false
              schema:
                example: Wed, 01 Dec 2021 09:23:06 GMT
                type: string
              style: simple
            Connection:
              explode: false
              schema:
                example: keep-alive
                type: string
              style: simple
            Keep-Alive:
              explode: false
              schema:
                example: timeout=5
                type: string
              style: simple
        "400":
          content:
            application/json:
              example:
                status: Bad Request
                message: title cannot be null
                data: {}
              schema:
                type: object
          description: Bad Request
          headers:
            X-Powered-By:
              explode: false
              schema:
                example: Express
                type: string
              style: simple
            Access-Control-Allow-Origin:
              explode: false
              schema:
                example: '*'
                type: string
              style: simple
            Content-Type:
              explode: false
              schema:
                example: application/json; charset=utf-8
                type: string
              style: simple
            Content-Length:
              explode: false
              schema:
                example: 67
                type: integer
              style: simple
            ETag:
              explode: false
              schema:
                example: W/"43-zSU6frTcl9gHoYQDTq06JCvAh0o"
                type: string
              style: simple
            Date:
              explode: false
              schema:
                example: Wed, 01 Dec 2021 09:23:19 GMT
                type: string
              style: simple
            Connection:
              explode: false
              schema:
                example: keep-alive
                type: string
              style: simple
            Keep-Alive:
              explode: false
              schema:
                example: timeout=5
                type: string
              style: simple
      summary: Create
      tags:
      - ACTIVITY
  /activity-groups/{id}:
    delete:
      operationId: deleteActivityById
      parameters:
      - description: The id of the activity to delete
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              example:
                status: Success
                message: Success
                data: {}
              schema:
                type: object
          description: OK
          headers:
            X-Powered-By:
              explode: false
              schema:
                example: Express
                type: string
              style: simple
            Access-Control-Allow-Origin:
              explode: false
              schema:
                example: '*'
                type: string
              style: simple
            Content-Type:
              explode: false
              schema:
                example: application/json; charset=utf-8
                type: string
              style: simple
            Content-Length:
              explode: false
              schema:
                example: 50
                type: integer
              style: simple
            ETag:
              explode: false
              schema:
                example: W/"32-xh+/pf6fWX8aREte3ywdfdCafcQ"
                type: string
              style: simple
            Date:
              explode: false
              schema:
                example: Wed, 01 Dec 2021 09:28:14 GMT
                type: string
              style: simple
            Connection:
              explode: false
              schema:
                example: keep-alive
                type: string
              style: simple
            Keep-Alive:
              explode: false
              schema:
                example: timeout=5
                type: string
              style: simple
        "404":
          content:
            application/json:
              example:
                status: Not Found
                message: Activity with ID 1381738 Not Found
                data: {}
              schema:
                type: object
          description: Not Found
          headers:
            X-Powered-By:
              explode: false
              schema:
                example: Express
                type: string
              style: simple
            Access-Control-Allow-Origin:
              explode: false
              schema:
                example: '*'
                type: string
              style: simple
            Content-Type:
              explode: false
              schema:
                example: application/json; charset=utf-8
                type: string
              style: simple
            Content-Length:
              explode: false
              schema:
                example: 79
                type: integer
              style: simple
            ETag:
              explode: false
              schema:
                example: W/"4f-qgnKSrigz/7Wmltk2sFU19H4QEk"
                type: string
              style: simple
            Date:
              explode: false
              schema:
                example: Wed, 01 Dec 2021 09:28:02 GMT
                type: string
              style: simple
            Connection:
              explode: false
              schema:
                example: keep-alive
                type: string
              style: simple
            Keep-Alive:
              explode: false
              schema:
                example: timeout=5
                type: string
              style: simple
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      summary: Delete
      tags:
      - ACTIVITY
    get:
      operationId: showActivityById
      parameters:
      - description: The id of the activity to retrieve
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              example:
                status: Success
                message: Success
                data:
                  id: 2
                  email: ad0286a7-bec4-405c-96e2-cd472c18a1e7@skyshi.com
                  title: coba 4
                  created_at: 2021-11-30T05:29:24.000Z
                  updated_at: 2021-11-30T05:29:24.000Z
                  deleted_at: null
              schema:
                type: object
          description: OK
          headers:
            X-Powered-By:
              explode: false
              schema:
                example: Express
                type: string
              style: simple
            Access-Control-Allow-Origin:
              explode: false
              schema:
                example: '*'
                type: string
              style: simple
            Content-Type:
              explode: false
              schema:
                example: application/json; charset=utf-8
                type: string
              style: simple
            Content-Length:
              explode: false
              schema:
                example: 229
                type: integer
              style: simple
            ETag:
              explode: false
              schema:
                example: W/"e5-m7C/ei1xhxZvcR7WPUB3REC4Rfo"
                type: string
              style: simple
            Date:
              explode: false
              schema:
                example: Wed, 01 Dec 2021 09:22:27 GMT
                type: string
              style: simple
            Connection:
              explode: false
              schema:
                example: keep-alive
                type: string
              style: simple
            Keep-Alive:
              explode: false
              schema:
                example: timeout=5
                type: string
              style: simple
        "404":
          content:
            application/json:
              example:
                status: Not Found
                message: Activity with ID 8328923 Not Found
                data: {}
              schema:
                type: object
          description: Not Found
          headers:
            X-Powered-By:
              explode: false
              schema:
                example: Express
                type: string
              style: simple
            Access-Control-Allow-Origin:
              explode: false
              schema:
                example: '*'
                type: string
              style: simple
            Content-Type:
              explode: false
              schema:
                example: application/json; charset=utf-8
                type: string
              style: simple
            Content-Length:
              explode: false
              schema:
                example: 79
                type: integer
              style: simple
            ETag:
              explode: false
              schema:
                example: W/"4f-X1tpbLO8bTFRVKfaxqNep2noZ2U"
                type: string
              style: simple
            Date:
              explode: false
              schema:
                example: Wed, 01 Dec 2021 09:22:47 GMT
                type: string
              style: simple
            Connection:
              explode: false
              schema:
                example: keep-alive
                type: string
              style: simple
            Keep-Alive:
              explode: false
              schema:
                example: timeout=5
                type: string
              style: simple
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      summary: Get One
      tags:
      - ACTIVITY
    patch:
      operationId: patchActivityById
      parameters:
      - description: The id of the activity to update
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              example:
                title: testing
              type: object
      responses:
        "200":
          content:
            application/json:
              example:
                status: Success
                message: Success
                data:
                  id: 2
                  email: ad0286a7-bec4-405c-96e2-cd472c18a1e7@skyshi.com
                  title: testing
                  created_at: 2021-11-30T05:29:24.000Z
                  updated_at: 2021-12-01T09:23:56.945Z
                  deleted_at: null
              schema:
                type: object
          description: OK
          headers:
            X-Powered-By:
              explode: false
              schema:
                example: Express
                type: string
              style: simple
            Access-Control-Allow-Origin:
              explode: false
              schema:
                example: '*'
                type: string
              style: simple
            Content-Type:
              explode: false
              schema:
                example: application/json; charset=utf-8
                type: string
              style: simple
            Content-Length:
              explode: false
              schema:
                example: 230
                type: integer
              style: simple
            ETag:
              explode: false
              schema:
                example: W/"e6-mqw3HnQfQVaOnHl8PkMk71MwTfg"
                type: string
              style: simple
            Date:
              explode: false
              schema:
                example: Wed, 01 Dec 2021 09:23:57 GMT
                type: string
              style: simple
            Connection:
              explode: false
              schema:
                example: keep-alive
                type: string
              style: simple
            Keep-Alive:
              explode: false
              schema:
                example: timeout=5
                type: string
              style: simple
        "400":
          content:
            application/json:
              example:
                status: Bad Request
                message: title cannot be null
                data: {}
              schema:
                type: object
          description: Bad Request
          headers:
            X-Powered-By:
              explode: false
              schema:
                example: Express
                type: string
              style: simple
            Access-Control-Allow-Origin:
              explode: false
              schema:
                example: '*'
                type: string
              style: simple
            Content-Type:
              explode: false
              schema:
                example: application/json; charset=utf-8
                type: string
              style: simple
            Content-Length:
              explode: false
              schema:
                example: 67
                type: integer
              style: simple
            ETag:
              explode: false
              schema:
                example: W/"43-zSU6frTcl9gHoYQDTq06JCvAh0o"
                type: string
              style: simple
            Date:
              explode: false
              schema:
                example: Wed, 01 Dec 2021 09:26:55 GMT
                type: string
              style: simple
            Connection:
              explode: false
              schema:
                example: keep-alive
                type: string
              style: simple
            Keep-Alive:
              explode: false
              schema:
                example: timeout=5
                type: string
              style: simple
        "404":
          content:
            application/json:
              example:
                status: Not Found
                message: Activity with ID 7329323 Not Found
                data: {}
              schema:
                type: object
          description: Not Found
          headers:
            X-Powered-By:
              explode: false
              schema:
                example: Express
                type: string
              style: simple
            Access-Control-Allow-Origin:
              explode: false
              schema:
                example: '*'
                type: string
              style: simple
            Content-Type:
              explode: false
              schema:
                example: application/json; charset=utf-8
                type: string
              style: simple
            Content-Length:
              explode: false
              schema:
                example: 79
                type: integer
              style: simple
            ETag:
              explode: false
              schema:
                example: W/"4f-O4NFMVLayndESb0e9APvjWGE/pU"
                type: string
              style: simple
            Date:
              explode: false
              schema:
                example: Wed, 01 Dec 2021 09:27:28 GMT
                type: string
              style: simple
            Connection:
              explode: false
              schema:
                example: keep-alive
                type: string
              style: simple
            Keep-Alive:
              explode: false
              schema:
                example: timeout=5
                type: string
              style: simple
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      summary: Update
      tags:
      - ACTIVITY
  /todo-items:
    get:
      parameters:
      - example: "2"
        explode: true
        in: query
        name: activity_group_id
        required: false
        schema:
          type: integer
        style: form
      responses:
        "200":
          content:
            application/json:
              examples:
                example-0:
                  summary: SUCCESS (without query)
                  value:
                    status: Success
                    message: Success
                    data:
                    - id: 1
                      activity_group_id: "2"
                      title: item1
                      is_active: "1"
                      priority: very-high
                      created_at: 2021-11-30T18:09:19.000Z
                      updated_at: 2021-11-30T20:23:12.000Z
                      deleted_at: null
                    - id: 3
                      activity_group_id: "1180"
                      title: item1
                      is_active: "1"
                      priority: very-high
                      created_at: 2021-11-30T20:34:56.000Z
                      updated_at: 2021-11-30T20:34:57.000Z
                      deleted_at: null
                example-1:
                  summary: SUCCESS (with query)
                  value:
                    status: Success
                    message: Success
                    data:
                    - id: 1
                      activity_group_id: "2"
                      title: item1
                      is_active: "1"
                      priority: very-high
                      created_at: 2021-11-30T18:09:19.000Z
                      updated_at: 2021-11-30T20:23:12.000Z
                      deleted_at: null
              schema:
                type: object
          description: OK
          headers:
            X-Powered-By:
              explode: false
              schema:
                example: Express
                type: string
              style: simple
            Access-Control-Allow-Origin:
              explode: false
              schema:
                example: '*'
                type: string
              style: simple
            Content-Type:
              explode: false
              schema:
                example: application/json; charset=utf-8
                type: string
              style: simple
            Content-Length:
              explode: false
              schema:
                example: 424
                type: integer
              style: simple
            ETag:
              explode: false
              schema:
                example: W/"1a8-LUXpko9eeImTS7l8RHbS2lSLw/Q"
                type: string
              style: simple
            Date:
              explode: false
              schema:
                example: Wed, 01 Dec 2021 09:30:43 GMT
                type: string
              style: simple
            Connection:
              explode: false
              schema:
                example: keep-alive
                type: string
              style: simple
            Keep-Alive:
              explode: false
              schema:
                example: timeout=5
                type: string
              style: simple
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      summary: Get All
      tags:
      - TODO
    post:
      requestBody:
        content:
          application/json:
            schema:
              example:
                activity_group_id: 3
                title: item 5.4
              type: object
      responses:
        "200":
          content:
            application/json: {}
          description: Successful response
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      summary: Create
      tags:
      - TODO
  /todo-items/{id}:
    delete:
      operationId: deleteToDoById
      parameters:
      - description: The id of the ToDo to delete
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json: {}
          description: Successful response
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      summary: Delete
      tags:
      - TODO
    get:
      operationId: showToDoById
      parameters:
      - description: The id of the ToDo to retrieve
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json: {}
          description: Successful response
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      summary: Get One
      tags:
      - TODO
    patch:
      operationId: patchToDoById
      parameters:
      - description: The id of the ToDo to update
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              example:
                title: testing
              type: object
      responses:
        "200":
          content:
            application/json: {}
          description: Successful response
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      summary: Update
      tags:
      - TODO
components:
  schemas:
    ActivityGroup:
      properties:
        id:
          format: int64
          type: integer
        title:
          type: string
        email:
          type: string
        created_at:
          type: string
        updated_at":
          type: string
        deleted_at":
          nullable: true
          type: string
      required:
      - title
      type: object
    ActivityGroups:
      items:
        $ref: '#/components/schemas/ActivityGroup'
      type: array
    ToDo:
      properties:
        id:
          format: int64
          type: integer
        activity_group_id:
          $ref: '#/components/schemas/ActivityGroup'
        title:
          type: string
        is_active:
          type: boolean
        priority:
          type: string
        created_at:
          type: string
        updated_at":
          type: string
        deleted_at":
          nullable: true
          type: string
      required:
      - title
      type: object
    ToDos:
      items:
        $ref: '#/components/schemas/ToDo'
      type: array
    Error:
      properties:
        code:
          format: int32
          type: integer
        message:
          type: string
      required:
      - code
      - message
      type: object